refactor: remove slice + move state to context
