+++ src/components/recovery/RecoveryDetails/index.tsx
@@ -12,7 +12,7 @@ import useSafeInfo from '@/hooks/useSafeInfo'
 import ErrorMessage from '@/components/tx/ErrorMessage'
 import { RecoverySigners } from '../RecoverySigners'
 import { Errors, logError } from '@/services/exceptions'
-import type { RecoveryQueueItem } from '@/store/recoverySlice'
+import type { RecoveryQueueItem } from '@/components/recovery/RecoveryLoaderContext'
 
 import txDetailsCss from '@/components/transactions/TxDetails/styles.module.css'
 import summaryCss from '@/components/transactions/TxDetails/Summary/styles.module.css'
