+++ src/components/recovery/RecoverySigners/index.tsx
@@ -8,7 +8,7 @@ import { Countdown } from '@/components/common/Countdown'
 import { ExecuteRecoveryButton } from '../ExecuteRecoveryButton'
 import { SkipRecoveryButton } from '../SkipRecoveryButton'
 import { useRecoveryTxState } from '@/hooks/useRecoveryTxState'
-import type { RecoveryQueueItem } from '@/store/recoverySlice'
+import type { RecoveryQueueItem } from '@/components/recovery/RecoveryLoaderContext'
 
 import txSignersCss from '@/components/transactions/TxSigners/styles.module.css'
 import { formatDate } from '@/utils/date'
