+++ src/components/settings/Recovery/index.tsx
@@ -8,8 +8,7 @@ import { Chip } from '@/components/common/Chip'
 import ExternalLink from '@/components/common/ExternalLink'
 import { DelayModifierRow } from './DelayModifierRow'
 import useIsSafeOwner from '@/hooks/useIsSafeOwner'
-import { useAppSelector } from '@/store'
-import { selectRecovery } from '@/store/recoverySlice'
+import { RecoveryLoaderContext } from '@/components/recovery/RecoveryLoaderContext'
 import EthHashInfo from '@/components/common/EthHashInfo'
 import EnhancedTable from '@/components/common/EnhancedTable'
 import InfoIcon from '@/public/images/notifications/info.svg'
@@ -69,11 +68,11 @@ const headCells = [
 
 export function Recovery(): ReactElement {
   const { setTxFlow } = useContext(TxModalContext)
-  const recovery = useAppSelector(selectRecovery)
+  const [recovery] = useContext(RecoveryLoaderContext).state
   const isOwner = useIsSafeOwner()
 
   const rows = useMemo(() => {
-    return recovery.flatMap((delayModifier) => {
+    return recovery?.flatMap((delayModifier) => {
       const { guardians, txCooldown, txExpiration } = delayModifier
 
       return guardians.map((guardian) => {
@@ -138,7 +137,7 @@ export function Recovery(): ReactElement {
             Enabling the Account recovery module will require a transactions.
           </Typography>
 
-          {recovery.length === 0 ? (
+          {recovery?.length === 0 ? (
             <>
               <Alert severity="info">
                 Unhappy with the provided option? {/* TODO: Add link */}
@@ -159,9 +158,9 @@ export function Recovery(): ReactElement {
                 )}
               </CheckWallet>
             </>
-          ) : (
+          ) : rows ? (
             <EnhancedTable rows={rows} headCells={headCells} />
-          )}
+          ) : null}
         </Grid>
       </Grid>
     </Paper>
