+++ src/hooks/useIsGuardian.ts
@@ -1,8 +1,11 @@
-import { useAppSelector } from '@/store'
-import { selectDelayModifierByGuardian } from '@/store/recoverySlice'
+import { useContext } from 'react'
+
+import { selectDelayModifierByGuardian } from '@/services/recovery/selectors'
 import useWallet from './wallets/useWallet'
+import { RecoveryLoaderContext } from '@/components/recovery/RecoveryLoaderContext'
 
 export function useIsGuardian() {
+  const [data] = useContext(RecoveryLoaderContext).state
   const wallet = useWallet()
-  return !!useAppSelector((state) => selectDelayModifierByGuardian(state, wallet?.address ?? ''))
+  return !wallet?.address || !data || !selectDelayModifierByGuardian(data, wallet.address)
 }
