+++ src/services/recovery/__tests__/selectors.test.ts
@@ -6,36 +6,28 @@ import {
   selectRecoveryQueues,
   selectDelayModifierByTxHash,
   selectDelayModifierByAddress,
-} from '../recoverySlice'
-import type { RecoveryState } from '../recoverySlice'
-import type { RootState } from '..'
+} from '../selectors'
+import type { RecoveryStateItem } from '@/components/recovery/RecoveryLoaderContext'
 
-describe('recoverySlice', () => {
+describe('selectors', () => {
   describe('selectDelayModifierByGuardian', () => {
     it('should return the Delay Modifier for the given guardian', () => {
       const delayModifier1 = {
         guardians: [faker.finance.ethereumAddress(), faker.finance.ethereumAddress()],
         queue: [{ timestamp: BigNumber.from(1) }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier2 = {
         guardians: [faker.finance.ethereumAddress()],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier3 = {
         guardians: [faker.finance.ethereumAddress()],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const data = [delayModifier1, delayModifier2, delayModifier3]
 
-      expect(
-        selectDelayModifierByGuardian(
-          {
-            recovery: { data },
-          } as unknown as RootState,
-          delayModifier1.guardians[0],
-        ),
-      ).toStrictEqual(delayModifier1)
+      expect(selectDelayModifierByGuardian(data, delayModifier1.guardians[0])).toStrictEqual(delayModifier1)
     })
   })
 
@@ -43,23 +35,19 @@ describe('recoverySlice', () => {
     it('should return all recovery queues sorted by timestamp', () => {
       const delayModifier1 = {
         queue: [{ timestamp: BigNumber.from(1) }, { timestamp: BigNumber.from(3) }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier2 = {
         queue: [{ timestamp: BigNumber.from(2) }, { timestamp: BigNumber.from(5) }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier3 = {
         queue: [{ timestamp: BigNumber.from(4) }, { timestamp: BigNumber.from(6) }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const data = [delayModifier1, delayModifier2, delayModifier3]
 
-      expect(
-        selectRecoveryQueues({
-          recovery: { data },
-        } as unknown as RootState),
-      ).toStrictEqual([
+      expect(selectRecoveryQueues(data)).toStrictEqual([
         { timestamp: BigNumber.from(1) },
         { timestamp: BigNumber.from(2) },
         { timestamp: BigNumber.from(3) },
@@ -76,53 +64,39 @@ describe('recoverySlice', () => {
 
       const delayModifier1 = {
         queue: [{ transactionHash: txHash }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier2 = {
         queue: [{ transactionHash: faker.string.hexadecimal() }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier3 = {
         queue: [{ transactionHash: faker.string.hexadecimal() }],
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const data = [delayModifier1, delayModifier2, delayModifier3]
 
-      expect(
-        selectDelayModifierByTxHash(
-          {
-            recovery: { data },
-          } as unknown as RootState,
-          txHash,
-        ),
-      ).toStrictEqual(delayModifier1)
+      expect(selectDelayModifierByTxHash(data, txHash)).toStrictEqual(delayModifier1)
     })
   })
 
   describe('selectDelayModifierByAddress', () => {
-    it('should return the Delay Modifier for the given txHash', () => {
+    it('should return the Delay Modifier for the given address', () => {
       const delayModifier1 = {
         address: faker.finance.ethereumAddress(),
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier2 = {
         address: faker.finance.ethereumAddress(),
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const delayModifier3 = {
         address: faker.finance.ethereumAddress(),
-      } as unknown as RecoveryState[number]
+      } as unknown as RecoveryStateItem
 
       const data = [delayModifier1, delayModifier2, delayModifier3]
 
-      expect(
-        selectDelayModifierByAddress(
-          {
-            recovery: { data },
-          } as unknown as RootState,
-          delayModifier2.address,
-        ),
-      ).toStrictEqual(delayModifier2)
+      expect(selectDelayModifierByAddress(data, delayModifier3.address)).toStrictEqual(delayModifier3)
     })
   })
 })
