+++ src/store/index.ts
@@ -29,7 +29,6 @@ import { safeAppsSlice } from './safeAppsSlice'
 import { safeMessagesListener, safeMessagesSlice } from './safeMessagesSlice'
 import { pendingSafeMessagesSlice } from './pendingSafeMessagesSlice'
 import { batchSlice } from './batchSlice'
-import { recoverySlice } from './recoverySlice'
 
 const rootReducer = combineReducers({
   [chainsSlice.name]: chainsSlice.reducer,
@@ -50,7 +49,6 @@ const rootReducer = combineReducers({
   [safeMessagesSlice.name]: safeMessagesSlice.reducer,
   [pendingSafeMessagesSlice.name]: pendingSafeMessagesSlice.reducer,
   [batchSlice.name]: batchSlice.reducer,
-  [recoverySlice.name]: recoverySlice.reducer,
 })
 
 const persistedSlices: (keyof PreloadedState<RootState>)[] = [
